Robustify the test for issue #12585
