+++ bokehjs/test/baselines/linux/Bug__in_issue_#12585__doesn't_allow_support_for_line_policy=none_with_mode=vline.blf
@@ -1,0 +1,4 @@
+Plot bbox=[0, 0, 200, 200]
+  CartesianFrame bbox=[29, 5, 166, 173]
+  LinearAxis bbox=[29, 178, 166, 22]
+  LinearAxis bbox=[0, 5, 29, 173]
