+++ bokehjs/test/interactive.ts
@@ -65,8 +65,8 @@ export class PlotActions {
     return this.target.canvas.events_el
   }
 
-  async hover(xy0: Point, xy1: Point, n?: number): Promise<void> {
-    await this.emit(this._hover({type: "line", xy0, xy1, n}))
+  async hover(xy0: Point, xy1?: Point, n?: number): Promise<void> {
+    await this.emit(this._hover({type: "line", xy0, xy1: xy1 ?? xy0, n}))
   }
 
   async move(xy0: Point, xy1: Point, n?: number, pressure?: number): Promise<void> {
